third_party/blink/web_tests/virtual/controls-refresh/datetimelocal-picker/datetimelocal-select-value.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="datetime-local" id="datetime-local" value="2019-02-14T13:02">
<script>

let t = async_test('Test select value in datetime-local popup');

function selectValue() {
  if (!internals.runtimeFlags.formControlsRefreshEnabled) {
    t.done();
  }

  let dateTimeElement = document.getElementById("datetime-local");
  dateTimeElement.addEventListener("change", t.step_func_done(() => {
    assert_equals(dateTimeElement.value, "2019-02-27T17:03");
  }));
  clickDayCellAt(3, 4);
  clickTimeCellAt(0, 4);
  clickTimeCellAt(1, 1);
  clickSubmitTimeButton();
}

t.step(() => {
  openPicker(document.getElementById('datetime-local'), t.step_func(selectValue));
});
</script>
</body>
</html>
